Implemented FileStorage class: deals saving tasks in the file
